+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * added tuple for session state cancellation
 * removed the requirement to send a page view before an event can be tracked
 * fixed collapsing sessions for statistics
+* fixed reading growth when there is no data
 
 ## 3.4.0